migrate to create react app
