--- .jshintrc
@@ -1,4 +1,0 @@
-{
-  "esversion": 6,
-  "-W138": true
-}
