+++ package.json
@@ -1,64 +1,49 @@
 {
   "name": "react-weather",
-  "version": "1.0.2",
-  "description": "Weather application for viewing weather",
+  "version": "2.0.0",
+  "description": "React application for viewing weather repots",
   "main": "index.js",
   "scripts": {
-    "clean": "rimraf ./build/*",
-    "jsonServer": "json-server --watch db.json --routes routes.json --port 3004 ",
-    "dev": "NODE_ENV=dev webpack-dev-server --progress",
-    "bundle:prod": "NODE_ENV=prod webpack -p --progress",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
+    "clean": "rimraf node_modules"
   },
   "author": "William S Fuller <wsfuller.dev@gmail.com>",
   "license": "ISC",
   "devDependencies": {
-    "babel": "~6.23.0",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-decorators-legacy": "~1.3.4",
-    "babel-preset-es2015": "~6.22.0",
-    "babel-preset-react": "~6.23.0",
-    "babel-preset-stage-0": "~6.22.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.5",
-    "eslint": "^4.5.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "favicons-webpack-plugin": "~0.0.7",
-    "file-loader": "~0.11.2",
-    "html-webpack-plugin": "~2.30.1",
-    "image-webpack-loader": "^3.3.1",
-    "node-sass": "^4.14.1",
-    "postcss-loader": "~1.3.3",
-    "postcss-smart-import": "~0.7.5",
-    "precss": "~2.0.0",
-    "react-addons-test-utils": "~15.6.0",
+    "react-scripts": "^3.4.1",
     "redux-devtools-extension": "~2.13.2",
-    "resolve-url-loader": "~2.1.0",
-    "rimraf": "~2.6.1",
-    "sass-loader": "~6.0.6",
-    "style-loader": "~0.18.2",
-    "url-loader": "~0.5.9",
-    "webpack": "~3.5.5",
-    "webpack-dev-server": "~2.7.1"
+    "rimraf": "^3.0.2"
   },
   "dependencies": {
+    "@fluentui/react": "^7.115.3",
     "axios": "^0.19.2",
-    "classnames": "^2.2.5",
-    "flexboxgrid": "^6.3.1",
+    "classnames": "^2.2.6",
     "google-map-react": "^0.24.0",
     "lodash": "^4.17.4",
-    "material-ui": "^0.19.0",
     "moment": "^2.18.1",
-    "normalize-scss": "^7.0.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-redux": "^5.0.6",
-    "react-router-dom": "^4.2.2",
+    "react-router-dom": "^5.2.0",
     "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.7.2",
     "redux-logger": "^3.0.6",
     "redux-promise-middleware": "^4.4.1",
     "redux-thunk": "^2.2.0"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
