--- postcss.config.js
@@ -1,19 +1,0 @@
-module.exports = {
-	plugins: [
-		require('postcss-smart-import')(
-			{
-				/* ...options */
-			}
-		),
-		require('precss')(
-			{
-				/* ...options */
-			}
-		),
-		require('autoprefixer')(
-			{
-				/* ...options */
-			}
-		)
-	]
-};
