+++ src/components/App/App.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+// import { connect } from 'react-redux';
+// import Layout from '../Layout.jsx';
+// import Config from '../../helpers/config.js';
+
+function App(){
+	return (
+		<div>React Weather Application</div>
+	)
+}
+
+// class App extends Component {
+// 	render() {
+// 		return <Layout app_name={Config.APPLICATION_NAME} />;
+// 	}
+// }
+
+
+export default App;
